// src/models/azure-client/iterations.d.ts
import { CollectionResult } from './common';

// Generated by https://quicktype.io

export type IterationsResult = CollectionResult<IterationInfo>

export interface IterationInfo {
	id:         string;
	name:       string;
	path:       string;
	attributes: Attributes;
	url:        string;
}

export interface Attributes {
	startDate:  null;
	finishDate: null;
	timeFrame:  string;
}